==> notebooks/01_ingesta.ipynb <==
# Importar librerías
from pyspark.sql import SparkSession
from pyspark.sql.functions import col, to_date, year, month
import os

spark = SparkSession.builder \
    .appName("SECOP_Ingesta") \
    .master("local[*]") \
    .config("spark.executor.memory", "8g") \
    .config("spark.driver.memory", "4g") \
    .getOrCreate()

print(f"Spark Version: {spark.version}")
print(f"Spark Master: {spark.sparkContext.master}")

import requests
import json


api_url = "https://www.datos.gov.co/resource/jbjy-vk9h.json?$query=SELECT%0A%20%20%60nombre_entidad%60%2C%0A%20%20%60nit_entidad%60%2C%0A%20%20%60departamento%60%2C%0A%20%20%60ciudad%60%2C%0A%20%20%60localizaci_n%60%2C%0A%20%20%60orden%60%2C%0A%20%20%60sector%60%2C%0A%20%20%60rama%60%2C%0A%20%20%60entidad_centralizada%60%2C%0A%20%20%60proceso_de_compra%60%2C%0A%20%20%60id_contrato%60%2C%0A%20%20%60referencia_del_contrato%60%2C%0A%20%20%60estado_contrato%60%2C%0A%20%20%60codigo_de_categoria_principal%60%2C%0A%20%20%60descripcion_del_proceso%60%2C%0A%20%20%60tipo_de_contrato%60%2C%0A%20%20%60modalidad_de_contratacion%60%2C%0A%20%20%60justificacion_modalidad_de%60%2C%0A%20%20%60fecha_de_firma%60%2C%0A%20%20%60fecha_de_inicio_del_contrato%60%2C%0A%20%20%60fecha_de_fin_del_contrato%60%2C%0A%20%20%60condiciones_de_entrega%60%2C%0A%20%20%60tipodocproveedor%60%2C%0A%20%20%60documento_proveedor%60%2C%0A%20%20%60proveedor_adjudicado%60%2C%0A%20%20%60es_grupo%60%2C%0A%20%20%60es_pyme%60%2C%0A%20%20%60habilita_pago_adelantado%60%2C%0A%20%20%60liquidaci_n%60%2C%0A%20%20%60obligaci_n_ambiental%60%2C%0A%20%20%60obligaciones_postconsumo%60%2C%0A%20%20%60reversion%60%2C%0A%20%20%60origen_de_los_recursos%60%2C%0A%20%20%60destino_gasto%60%2C%0A%20%20%60valor_del_contrato%60%2C%0A%20%20%60valor_de_pago_adelantado%60%2C%0A%20%20%60valor_facturado%60%2C%0A%20%20%60valor_pendiente_de_pago%60%2C%0A%20%20%60valor_pagado%60%2C%0A%20%20%60valor_amortizado%60%2C%0A%20%20%60valor_pendiente_de%60%2C%0A%20%20%60valor_pendiente_de_ejecucion%60%2C%0A%20%20%60estado_bpin%60%2C%0A%20%20%60c_digo_bpin%60%2C%0A%20%20%60anno_bpin%60%2C%0A%20%20%60saldo_cdp%60%2C%0A%20%20%60saldo_vigencia%60%2C%0A%20%20%60espostconflicto%60%2C%0A%20%20%60dias_adicionados%60%2C%0A%20%20%60puntos_del_acuerdo%60%2C%0A%20%20%60pilares_del_acuerdo%60%2C%0A%20%20%60urlproceso%60%2C%0A%20%20%60nombre_representante_legal%60%2C%0A%20%20%60nacionalidad_representante_legal%60%2C%0A%20%20%60domicilio_representante_legal%60%2C%0A%20%20%60tipo_de_identificaci_n_representante_legal%60%2C%0A%20%20%60identificaci_n_representante_legal%60%2C%0A%20%20%60g_nero_representante_legal%60%2C%0A%20%20%60presupuesto_general_de_la_nacion_pgn%60%2C%0A%20%20%60sistema_general_de_participaciones%60%2C%0A%20%20%60sistema_general_de_regal_as%60%2C%0A%20%20%60recursos_propios_alcald_as_gobernaciones_y_resguardos_ind_genas_%60%2C%0A%20%20%60recursos_de_credito%60%2C%0A%20%20%60recursos_propios%60%2C%0A%20%20%60ultima_actualizacion%60%2C%0A%20%20%60codigo_entidad%60%2C%0A%20%20%60codigo_proveedor%60%2C%0A%20%20%60fecha_inicio_liquidacion%60%2C%0A%20%20%60fecha_fin_liquidacion%60%2C%0A%20%20%60objeto_del_contrato%60%2C%0A%20%20%60duraci_n_del_contrato%60%2C%0A%20%20%60nombre_del_banco%60%2C%0A%20%20%60tipo_de_cuenta%60%2C%0A%20%20%60n_mero_de_cuenta%60%2C%0A%20%20%60el_contrato_puede_ser_prorrogado%60%2C%0A%20%20%60fecha_de_notificaci_n_de_prorrogaci_n%60%2C%0A%20%20%60nombre_ordenador_del_gasto%60%2C%0A%20%20%60tipo_de_documento_ordenador_del_gasto%60%2C%0A%20%20%60n_mero_de_documento_ordenador_del_gasto%60%2C%0A%20%20%60nombre_supervisor%60%2C%0A%20%20%60tipo_de_documento_supervisor%60%2C%0A%20%20%60n_mero_de_documento_supervisor%60%2C%0A%20%20%60nombre_ordenador_de_pago%60%2C%0A%20%20%60tipo_de_documento_ordenador_de_pago%60%2C%0A%20%20%60n_mero_de_documento_ordenador_de_pago%60%2C%0A%20%20%60documentos_tipo%60%2C%0A%20%20%60descripcion_documentos_tipo%60%0AWHERE%0A%20%20caseless_one_of(%60departamento%60%2C%20%22Distrito%20Capital%20de%20Bogot%C3%A1%22)%0A%20%20AND%20%60fecha_de_firma%60%0A%20%20%20%20%20%20%20%20BETWEEN%20%222025-10-01T19%3A14%3A39%22%20%3A%3A%20floating_timestamp%0A%20%20%20%20%20%20%20%20AND%20%222025-12-31T23%3A59%3A00%22%20%3A%3A%20floating_timestamp"

response = requests.get(api_url)
data = response.json()

pip install sodapy

from sodapy import Socrata

client = Socrata("www.datos.gov.co", None)

results = client.get(
    "jbjy-vk9h",
    query="""
    SELECT *
    WHERE 
        departamento = "Distrito Capital de Bogotá"
    AND
        fecha_de_firma >= '2025-01-01T00:00:00'
    AND
        fecha_de_firma < '2025-04-01T00:00:00'
    LIMIT 150000
    """
)

len(results)

import pandas as pd
import os

DATA_PATH = "/opt/spark-data"

os.makedirs(DATA_PATH, exist_ok=True)

df = pd.DataFrame.from_records(results)

df.to_csv(
    os.path.join(DATA_PATH, "SECOP_II_Contratos_Electronicos_Q1_2025.csv"),
    index=False,
    encoding="utf-8"
)

print("CSV guardado correctamente en /opt/spark-data")


# Guardar JSON localmente
json_path = "/opt/spark-data/raw/secop_contratos.json"
os.makedirs(os.path.dirname(json_path), exist_ok=True)

with open(json_path, 'w', encoding='utf-8') as f:
    for record in results:
        f.write(json.dumps(record, ensure_ascii=False) + '\n')

print(f"Datos guardados en: {json_path}")


print("Leyendo datos desde JSON...")
df_raw = spark.read.json(json_path)

print(f"Total de registros: {df_raw.count()}")
print(f"Total de columnas: {len(df_raw.columns)}")

df_raw.printSchema()

df_raw.show()

spark
df_spark = spark.createDataFrame(df)


len(df_spark.columns)

df_spark.printSchema()

# ==========================================================
# 1. CREAR SPARK SESSION
# ==========================================================

from pyspark.sql import SparkSession

spark = SparkSession.builder \
    .appName("SECOP_Ingesta") \
    .master("local[*]") \
    .getOrCreate()

print("Spark Version:", spark.version)


# ==========================================================
# 2. LEER JSON DESDE RAW
# ==========================================================

raw_path = "/opt/spark-data/raw/secop_contratos.json"

df_raw = spark.read.json(raw_path)

print("Registros cargados:", df_raw.count())
print("Columnas:", len(df_raw.columns))

df_raw.printSchema()
df_raw.show(5)


# ==========================================================
# 3. GUARDAR EN ZONA BRONZE (PARQUET)
# ==========================================================

bronze_path = "/opt/spark-data/bronze/secop_contratos"

import os
os.makedirs("/opt/spark-data/bronze", exist_ok=True)

# Reducimos particiones (evita problemas de memoria)
df_raw = df_raw.repartition(4)

df_raw.write \
    .mode("overwrite") \
    .parquet(bronze_path)

print("Datos guardados en Bronze:", bronze_path)


# ==========================================================
# 4. VERIFICACIÓN FINAL
# ==========================================================

df_check = spark.read.parquet(bronze_path)

print("Registros en Bronze:", df_check.count())
