==> notebooks/04_transformation.ipynb <==
# %%
from pyspark.sql import SparkSession
from pyspark.ml.feature import StandardScaler, PCA, VectorAssembler
from pyspark.ml import Pipeline, PipelineModel
from pyspark.sql.functions import col

# %%
spark = SparkSession.builder \
    .appName("SECOP_Transformaciones") \
    .master("local[*]") \
    .getOrCreate()

# Cargar datos transformados del notebook anterior
df = spark.read.parquet("/opt/spark-data/processed/secop_features.parquet")
print(f"Registros: {df.count():,}")
print(f"Columnas: {len(df.columns)}")

df.show()

scaler = StandardScaler(inputCol="features_raw",
                       outputCol = "features_scaled",
                       withMean=False,
                       withStd=True)

print("Aplicando StandardScaler")
scaler_model = scaler.fit(df)
df_scaled=scaler_model.transform(df)

df_scaled.show()

df_scaled.select("features_scaled").first()[0]

k=25

pca= PCA( k=k,
         inputCol="features_scaled",
         outputCol="features_pca")

print("Aplicando PCA")

pca_model = pca.fit(df_scaled)
df_pca = pca_model.transform(df_scaled)

df_pca.select("features_pca").first()[0]

explained_variance = pca_model.explainedVariance
sum(explained_variance)

pipeline_transform = Pipeline(stages=[scaler, pca])
pipeline_transform_model=pipeline_transform.fit(df)
df_final=pipeline_transform_model.transform(df)

df_ml=df_final.select("features_pca","valor_del_contrato_num")



pipeline_path= "/opt/spark-data/processed/transformation_pipeline"
output_path= "/opt/spark-data/processed/secop_ml_ready.parquet"
df_ml.write.mode("overwrite").parquet(output_path)
pipeline_transform_model.save(pipeline_path)
